+++ src/jawl_voicecompanion/__init__.py
@@ -18,6 +18,7 @@ from .voicemem_client import VoiceMemProcessClient, VoiceMemUnavailable
 from .jawl_web import JawlWebAdapter, JawlWebChatAdapter, JawlWebUnavailable
 from .ambient_memory import AmbientMemoryBuffer
 from .ambient_triage import AmbientTriageProvider, AmbientTriageUnavailable, OllamaTriageProvider
+from .ambient_audio import AmbientAudioASRBridge
 from .system_audio import SystemAudioLoopback, SystemAudioUnavailable
 
 __all__ = [
@@ -52,6 +53,7 @@ __all__ = [
     "AmbientTriageProvider",
     "AmbientTriageUnavailable",
     "OllamaTriageProvider",
+    "AmbientAudioASRBridge",
     "SystemAudioLoopback",
     "SystemAudioUnavailable",
 ]